feat：修正类名、字段，门店添加新的字段，修复添加权限时入参错误问题

+++ src/main/java/com/basic/domain/entity/DineAttributeOption.java
@@ -37,7 +37,7 @@ public class DineAttributeOption extends BasicEntity {
     /**
      * 价格调整(分)，可为负数
      */
-    private Integer priceAdjustment;
+    private Long priceAdjustment;
 
     /**
      * 排序
