+++ src/main/java/com/basic/domain/entity/DineDishAttributeOption.java
@@ -41,6 +41,6 @@ public class DineDishAttributeOption extends BasicEntity {
     /**
      * 覆盖价格调整，NULL则使用选项默认值
      */
-    private Integer priceAdjustmentOverride;
+    private Long priceAdjustmentOverride;
 
 }
